clean up and fix tests
